pyproject.toml
[project]
name = "tts-project"
version = "0.1.0"
description = "DOCX to MP3 Converter with Kokoro & Google Cloud TTS"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-cloud-texttospeech",
    "kokoro",
    "numpy",
    "python-docx",
    "pyyaml",
    "soundfile",
]
